+++ src/compare.ts
